async: allow passing of initfile to child process (FR#39)
Sometimes passing all your state setup in the async elisp call can be
painful, especially if the library is being used by another one (in my
case ob-async for org-mode). Luckily we have a method for dealing with
boilerplate setup called an initfile. This allows an initfile to be
specified for all async calls. It does not re-use the existing
initfile as the interactive one is likely too cluttered for a child
process.