Skip async for fast operations
Starting an async process for fast operations is clunky, so add a new
option/feature for skipping async if the operation is fast.  Same
device renames are fast (since it's just a directory entry change) and
renames and copys of small files are fast (arbitrarily picking 5 MB as
the cutoff for modern computers/disks, can be configured).