Preserve value of async-prompt-for-password when async-start is called
This allows callers of async-start to let-bind
async-prompt-for-password (e.g., to nil to prevent any prompting on
data that might contain arbitrary sexps).  The value in effect while
async-start is called is saved in the process's buffer, and is
therefore used by async-read-from-client since it makes the process's
buffer current.