Add variable to set process-query-on-exit-flag
This can't reliably be done by the caller of async-start, because
make-process will create an extra process for the stderr buffer that
cannot be easily found based on the returned process from async-start.

The new variable, async-process-noquery-on-exit, can be let-bound to
t around a call to async-start in order to cause the parent Emacs to
silently kill the child Emacs if the user quits the parent.

This can be useful for async processes run in timers for things like
notifications, where the user doesn't really care if a running process
actually finishes, since they're killing Emacs.