Use temporary file for async byte compile log
This is to avoid log file contention when multiple packages are compiled in
rapid succession.

Update the default value `async-byte-compile-log-file`. The
`temporary-file-directory` (implicitly) is used as a location where to store
temporary compilation logs. Update to the semantic is done with backward
compatibility in mind, that is to respect the directory part should it be set
by user.

Add `async-bytecomp--comp-buffer-to-file` macro to be used (`macroexpand`ed) in
`async-byte-recompile-directory` and `async-byte-compile-file`. Pass value
returned from expanded macro execution in child processes to
`async-bytecomp--file-to-comp-buffer` call-back.